requirements.txt
# requirements.txt
fastapi[standard]==0.128.0
uvicorn[standard]==0.40.0
sqlalchemy==2.0.45
alembic==1.17.2
pydantic==2.12.5
pydantic-settings==2.12.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.21
aiomysql==0.2.0
greenlet==3.3.0
python-dotenv==1.1.0
aiofiles==24.1.0

# Development
pytest==9.0.2
pytest-asyncio==1.3.0
httpx==0.28.1